Вынес связи в приложение Students
