+++ career_tracker_backend/specialties/models.py
@@ -68,45 +68,6 @@ class Course(models.Model):
         return self.name
 
 
-class StudentCourse(models.Model):
-    """Связь студентов с курсами"""
-
-    student = models.ForeignKey(
-        Student,
-        related_name='student_courses',
-        verbose_name='Студент',
-        on_delete=models.CASCADE
-    )
-    course = models.ForeignKey(
-        Course,
-        related_name='course_students',
-        verbose_name='Курс',
-        on_delete=models.CASCADE
-    )
-    status_payment = models.BooleanField(
-        verbose_name='Статус покупки',
-        default=False
-    )
-    status_course = models.BooleanField(
-        verbose_name='Статус изучения',
-        default=False
-    )
-
-    class Meta:
-        ordering = ('student',)
-        verbose_name = 'Связь студента и курса'
-        verbose_name_plural = 'Связь студентов и курсов'
-        constraints = [
-            models.UniqueConstraint(
-                fields=('student', 'course',),
-                name='unique_student_course'
-            )
-        ]
-
-    def __str__(self):
-        return f'студент {self.student} проходит курс {self.course}'
-
-
 class Grade(models.Model):
     '''Модель уровня знаний по специальности'''
 
@@ -175,33 +136,6 @@ class Skill(models.Model):
         return self.name
 
 
-class SkillStudent(models.Model):
-    student = models.ForeignKey(
-        Student,
-        on_delete=models.CASCADE,
-        related_name='skills_student',
-        verbose_name='Студент'
-    )
-    skill = models.ForeignKey(
-        Skill,
-        on_delete=models.CASCADE,
-        related_name='students_skill',
-        verbose_name='Навык'
-    )
-    status = models.BooleanField(
-        default=False,
-        verbose_name='Статус получения навыка'
-    )
-
-    class Meta:
-        ordering = ('student',)
-        verbose_name = 'Связь студента и навыка'
-        verbose_name_plural = 'Связь студентов и навыков'
-
-    def __str__(self):
-        return f'Студент {self.student} изучает навык {self.skill}'
-
-
 class Direction(models.Model):
     name = models.CharField(
         max_length=MAX_LENGHT,
@@ -246,30 +180,3 @@ class Sprint(models.Model):
 
     def __str__(self):
         return self.name
-
-
-class SprintStudent(models.Model):
-    sprint = models.ForeignKey(
-        Sprint,
-        on_delete=models.CASCADE,
-        related_name='students_sprint',
-        verbose_name='Спринт'
-    )
-    student = models.ForeignKey(
-        Student,
-        on_delete=models.CASCADE,
-        related_name='sprints_student',
-        verbose_name='Студент'
-    )
-    status = models.BooleanField(
-        default=True,
-        verbose_name='Статус прохождения спринта'
-    )
-
-    class Meta:
-        ordering = ('student',)
-        verbose_name = 'Связь студента и спринта'
-        verbose_name_plural = 'Связь студентов и спринтов'
-
-    def __str__(self):
-        return f'Студент {self.student} проходит спринт {self.sprint}'
